refactor: widget
